refactor: Migrate database primary and foreign keys from Int to Long
- Update all Room database entities (`Venue`, `Building`, `Floor`, `Node`, `Edge`, `Event`, `Vendor`, `Booth`, `FloorConnection`) to use `Long` for their primary keys (`id`) and foreign keys (`*_id`) instead of `Int`. This change provides a larger range for identifiers, preventing potential overflow issues as the database grows.

- Adjust DAOs (`VenueDao`, `FloorDao`, `EdgeDao`, etc.) and Repository interfaces/implementations to accept and query by `Long` instead of `Int`.

- Update `NodeMapViewModel` to use `Long` for `floorId` when loading data (floors, nodes, edges, connections), ensuring consistency with the updated database schema.

- Correct the return type of `insertVenue` in `VenueDao` from `Int` to `Long` in the `createVenueWithDefaults` transaction to properly handle the new key type.

- Remove an unused and outdated `core-ktx` test dependency from the `libs.versions.toml` file.

+++ app/src/main/java/ie/app/minimap/NodeMapViewModel.kt
@@ -41,7 +41,7 @@ class NodeMapViewModel @Inject constructor(
 
     // ---------------- FLOOR ----------------------
 
-    fun loadFloor(floorId: Int) {
+    fun loadFloor(floorId: Long) {
         viewModelScope.launch {
             floorRepository.getFloorById(floorId)
                 .collect { floor ->
@@ -56,7 +56,7 @@ class NodeMapViewModel @Inject constructor(
 
     // ---------------- NODE ----------------------
 
-    fun loadNodes(floorId: Int) {
+    fun loadNodes(floorId: Long) {
         viewModelScope.launch {
             nodeRepository.getNodeByFloor(floorId)
                 .collect { _nodes.value = it }
@@ -77,7 +77,7 @@ class NodeMapViewModel @Inject constructor(
 
     // ---------------- EDGE ----------------------
 
-    fun loadEdges(floorId: Int) {
+    fun loadEdges(floorId: Long) {
         viewModelScope.launch {
             edgeRepository.getEdgesByFloor(floorId)
                 .collect { _edges.value = it }
@@ -99,7 +99,7 @@ class NodeMapViewModel @Inject constructor(
 
     // ---------------- FLOOR CONNECTION ----------------------
 
-    fun loadFloorConnections(floorId: Int) {
+    fun loadFloorConnections(floorId: Long) {
         viewModelScope.launch {
             floorConnectionRepository.getConnectionsByFloor(floorId)
                 .collect { _connections.value = it }
