+++ app/src/main/java/ie/app/minimap/data/local/entity/Building.kt
@@ -19,9 +19,9 @@ import androidx.room.PrimaryKey
 )
 data class Building(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     @ColumnInfo(name = "venue_id")
-    val venueId: Int = 0,
+    val venueId: Long = 0,
     val name: String = "",
     val description: String = "",
     @ColumnInfo(name = "created_at")
